Flytta weight från Product till respektive subklasser pga oändliga problem. Uppdaterade shippable till double weight()

+++ src/main/java/com/example/Shippable.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public interface Shippable {
 
-    BigDecimal weight();
+    double weight();
     BigDecimal calculateShippingCost();
 
 }
